LRCI-3772 Move the event controllers together into the same package
