+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/handler/BaseEventHandler.java
@@ -9,11 +9,11 @@ import com.liferay.jethr0.bui1d.queue.BuildQueue;
 import com.liferay.jethr0.bui1d.repository.BuildEntityRepository;
 import com.liferay.jethr0.bui1d.repository.BuildParameterEntityRepository;
 import com.liferay.jethr0.bui1d.repository.BuildRunEntityRepository;
+import com.liferay.jethr0.event.controller.EventJmsController;
 import com.liferay.jethr0.jenkins.JenkinsQueue;
 import com.liferay.jethr0.jenkins.repository.JenkinsCohortEntityRepository;
 import com.liferay.jethr0.jenkins.repository.JenkinsNodeEntityRepository;
 import com.liferay.jethr0.jenkins.repository.JenkinsServerEntityRepository;
-import com.liferay.jethr0.jms.JMSEventHandler;
 import com.liferay.jethr0.project.repository.ProjectEntityRepository;
 
 import org.json.JSONObject;
@@ -46,6 +46,10 @@ public abstract class BaseEventHandler implements EventHandler {
 		return _eventHandlerContext.getBuildRunRepository();
 	}
 
+	protected EventJmsController getEventJmsController() {
+		return _eventHandlerContext.getEventJmsController();
+	}
+
 	protected JenkinsCohortEntityRepository getJenkinsCohortEntityRepository() {
 		return _eventHandlerContext.getJenkinsCohortEntityRepository();
 	}
@@ -60,10 +64,6 @@ public abstract class BaseEventHandler implements EventHandler {
 
 	protected JenkinsServerEntityRepository getJenkinsServerEntityRepository() {
 		return _eventHandlerContext.getJenkinsServerEntityRepository();
-	}
-
-	protected JMSEventHandler getJMSEventHandler() {
-		return _eventHandlerContext.getJMSEventHandler();
 	}
 
 	protected JSONObject getMessageJSONObject() {
