+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/handler/EventHandlerContext.java
@@ -9,11 +9,11 @@ import com.liferay.jethr0.bui1d.queue.BuildQueue;
 import com.liferay.jethr0.bui1d.repository.BuildEntityRepository;
 import com.liferay.jethr0.bui1d.repository.BuildParameterEntityRepository;
 import com.liferay.jethr0.bui1d.repository.BuildRunEntityRepository;
+import com.liferay.jethr0.event.controller.EventJmsController;
 import com.liferay.jethr0.jenkins.JenkinsQueue;
 import com.liferay.jethr0.jenkins.repository.JenkinsCohortEntityRepository;
 import com.liferay.jethr0.jenkins.repository.JenkinsNodeEntityRepository;
 import com.liferay.jethr0.jenkins.repository.JenkinsServerEntityRepository;
-import com.liferay.jethr0.jms.JMSEventHandler;
 import com.liferay.jethr0.project.repository.ProjectEntityRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -41,6 +41,10 @@ public class EventHandlerContext {
 		return _buildRunEntityRepository;
 	}
 
+	public EventJmsController getEventJmsController() {
+		return _eventJmsController;
+	}
+
 	public JenkinsCohortEntityRepository getJenkinsCohortEntityRepository() {
 		return _jenkinsCohortEntityRepository;
 	}
@@ -57,16 +61,12 @@ public class EventHandlerContext {
 		return _jenkinsServerEntityRepository;
 	}
 
-	public JMSEventHandler getJMSEventHandler() {
-		return _jmsEventHandler;
-	}
-
 	public ProjectEntityRepository getProjectEntityRepository() {
 		return _projectEntityRepository;
 	}
 
-	public void setJMSEventHandler(JMSEventHandler jmsEventHandler) {
-		_jmsEventHandler = jmsEventHandler;
+	public void setEventJmsController(EventJmsController eventJmsController) {
+		_eventJmsController = eventJmsController;
 	}
 
 	@Autowired
@@ -81,6 +81,8 @@ public class EventHandlerContext {
 	@Autowired
 	private BuildRunEntityRepository _buildRunEntityRepository;
 
+	private EventJmsController _eventJmsController;
+
 	@Autowired
 	private JenkinsCohortEntityRepository _jenkinsCohortEntityRepository;
 
@@ -92,8 +94,6 @@ public class EventHandlerContext {
 
 	@Autowired
 	private JenkinsServerEntityRepository _jenkinsServerEntityRepository;
-
-	private JMSEventHandler _jmsEventHandler;
 
 	@Autowired
 	private ProjectEntityRepository _projectEntityRepository;
